avcodec/ccaption_dec: Disable skipping of leading space
EIA-608 and SCC define explicit markup & placement.  It should not be up to the decoder to skip leading spaces.  If a subsequent subtitle format (ass, srt or vtt) has a preference or constraint to skip leading spaces, that should be done in the subtitle encoder, not the EIA-608/SCC decoder.